requirements.txt
Appium-Python-Client==2.9.0
selenium==4.9.1
urllib3<2
pytest==7.3.1
pytest-html==3.2.0
google-api-python-client==2.86.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
pillow==9.4.0
python-dotenv